Refactor ChatViewModel to use EncryptionService instance
